requirements.txt
attrs==17.3.0
bcrypt==3.1.4
cassandra-driver==3.12.0
cffi==1.11.2
colander==1.4
cornice==3.0.0
decorator==4.1.2
hupper==1.0
ipython==6.2.1
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.11.0
logmatic-python==0.1.7
parso==0.1.0
Paste==2.0.3
PasteDeploy==1.5.2
PasteScript==2.0.2
pexpect==4.3.0
pickleshare==0.7.4
plaster==1.0
plaster-pastedeploy==0.4.2
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
pycparser==2.18
Pygments==2.2.0
pyramid==1.9.1
pyramid-multiauth==0.9.0
pytest==3.3.1
python-json-logger==0.1.8
repoze.lru==0.7
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
traitlets==4.3.2
translationstring==1.3
venusian==1.1.0
waitress==1.1.0
wcwidth==0.1.7
WebOb==1.7.4
zope.deprecation==4.3.0
zope.interface==4.4.3
